% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_ids.R
\name{add_ids}
\alias{add_ids}
\title{Adds an id column to a given variable.}
\usage{
add_ids(df, idname = "id")
}
\arguments{
\item{df}{dataframe.  The data to add an id var to.}

\item{idname}{character. The name of the variable in the dataframe you want to be the id column.}
}
\description{
Adds an id column to a given variable.
}
