% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/add_prefix_to_table_names.R
\name{add_prefix_to_table_names}
\alias{add_prefix_to_table_names}
\title{Gets the vector of a particular variable within a dplyr postgres table.}
\usage{
add_prefix_to_table_names(df, prefix, except = NULL)
}
\arguments{
\item{pt}{postgres.  The postgres table.}

\item{variable}{character. The name of the variable to gather.}
}
\description{
Gets the vector of a particular variable within a dplyr postgres table.
}

