% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_over_vars.R
\name{apply_over_vars}
\alias{apply_over_vars}
\title{Apply a function over variables in a dataframe.}
\usage{
apply_over_vars(df, vars, f)
}
\arguments{
\item{df}{dataframe. The dataframe to apply the function over.}

\item{f}{function. The function to apply.}

\item{vars.}{A list of the variable names to apply to, or a character regex pattern to search for (using `get_vars`).}
}
\description{
Apply a function over variables in a dataframe.
}
