% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breakdown.R
\name{breakdown}
\alias{breakdown}
\title{Breakdown values of a variable by the number of people who have that value or a higher value.}
\usage{
breakdown(df, variable, breakdowns)
}
\arguments{
\item{df}{dataframe. The dataframe to analyze.}

\item{variable}{character. The value to breakdown.}

\item{breakdowns}{integer. A vector of values to breakdown.  Usually produced by seq(), but could be a vector.

e.g., \code{breakdown(data, 'age', seq(5, 90, by = 5))} will tell me how many people
are age 5 or older, how many people are age 10 or older, how many people are
age 15 or older, etc. to age 90.}
}
\description{
Breakdown values of a variable by the number of people who have that value or a higher value.
}

