% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chisq_test.R
\name{chisq_test}
\alias{chisq_test}
\title{Reformat stats::chisq.test to accept inputs as two vectors.}
\usage{
chisq_test(vector1, vector2)
}
\arguments{
\item{vector1.}{The first vector to compare.}

\item{vector2.}{The other vector to compare.}
}
\value{
chi square test on the joint frequency table of the two vectors.
}
\description{
Reformat stats::chisq.test to accept inputs as two vectors.
}
