% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/comparison_table.R
\name{comparison_table}
\alias{comparison_table}
\title{Compares a variable across a group, both visually and with an appropriate statistical test.}
\usage{
comparison_table(data = NULL, variable, groupby, type)
}
\arguments{
\item{data}{data frame. The data to test.}

\item{groupby}{character. If character, the name of the variable to group `variable` by.  Or you can pass a specific vector instead.}

\item{type}{character. "Continuous" if `variable` is continuous data (like age), or "categorical" if `variable` is categorical data (like hometown).}

\item{variable.}{If character, the name of the variable to test.  Or you can pass a specific vector instead.}
}
\description{
Compares a variable across a group, both visually and with an appropriate statistical test.
}

