% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/count_vars.R
\name{count_vars}
\alias{count_vars}
\title{Count the amount of "yes" responses that respondants have within a group of vars.}
\usage{
count_vars(df, vars, idname = "id", yes = 1, colname = "value")
}
\arguments{
\item{df}{dataframe.  The data to be transformed.}

\item{vars}{character. A vector of variables to check.}

\item{idname}{character. The name of the variable in the dataframe where respondent ids are kept.  Defaults to 'id'.}

\item{yes}{character. The value of the response counted as a "yes" response that you seek to tally.  Defaults to 1.}

\item{colname}{character. The output will be a two-column dataframe with ids and values. The names are the idname for ids and this (default 'value') for the values.}
}
\description{
Count the amount of "yes" responses that respondants have within a group of vars.
}

