% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut.R
\name{cut3}
\alias{cut3}
\title{Split a vector into bins.}
\usage{
cut3(x, cuts)
}
\arguments{
\item{x}{vector. The vector to split into bins.}

\item{cuts}{numeric. The number of bins to split.}
}
\description{
Split a vector into bins.
}
