% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_summary.R
\name{data_summary}
\alias{data_summary}
\title{Prints a summary of the dataframe.}
\usage{
data_summary(df, ...)
}
\arguments{
\item{df}{dataframe. The dataframe to summarize.}

\item{...}{list. Additional args to pass to var_summary.}
}
\description{
Prints a summary of the dataframe.
}
