% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/dplyr_table.R
\name{dplyr_table}
\alias{dplyr_table}
\title{Makes a fancy table, using dplyr-like syntax, but inspired from `tab` from STATA.
Try to make Hadley proud.}
\usage{
dplyr_table(.data, ..., freq = TRUE, percent = FALSE, byrow = TRUE,
  sort = TRUE, sort.decreasing = TRUE)
}
\arguments{
\item{.data}{data frame. The data to put in the table.}

\item{...}{The variables to plot in the table. Currently supports one or two variables.}

\item{freq}{logical. If true, frequencies are printed in the table.}

\item{percent}{logical. If true, percentages are printed in the table.}

\item{byrow}{logical. If true, percentages are row percentages.  If false, column percentages.  If NULL, individual cell percentages.}

\item{sort}{logical. If true, the resulting table is sorted (if possible).}

\item{sort.decreasing}{logical. If true, the sort is preformed in a descending manner.}
}
\description{
Makes a fancy table, using dplyr-like syntax, but inspired from `tab` from STATA.
Try to make Hadley proud.
}

