% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_na_cols.R, R/na.rm.R
\name{drop_na_cols}
\alias{drop_na_cols}
\alias{drop_na_cols}
\title{Drops columns from the dataframe that meet a certain threshold of NAs.}
\usage{
drop_na_cols(df)

drop_na_cols(df)
}
\arguments{
\item{df}{dataframe.  The dataframe to drop from.}

\item{treshold}{integer.  If a column has this many percent of NAs or more, drop.}

\item{df}{data.frame. The dataframe to drop from.}
}
\description{
Drops columns from the dataframe that meet a certain threshold of NAs.

Drops columns that are entirely NA from the dataframe.
}
