% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_values.R
\name{drop_values}
\alias{drop_values}
\title{Drop values from a variable.}
\usage{
drop_values(x, values, to_na = TRUE)
}
\arguments{
\item{x}{vector. The vector to drop values from.}

\item{values}{list. A list of values to drop.}

\item{to_na}{logical. If to_na is FALSE, values will be fully dropped instead of converted to NA.}
}
\description{
Drop values from a variable.
}
