% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/fetch_percent_table.R
\name{fetch_percent_table}
\alias{fetch_percent_table}
\title{Runs percent_table for melted data frames.}
\usage{
fetch_percent_table(var, data = get_data(), round = NULL)
}
\arguments{
\item{var}{characer.  The name of the variable you want percent breakdown for.}

\item{round}{integer. How many digits you want to round percentages to. Default no rounding.}

\item{data.}{The data to grap values from. This is assumed to be a three column
melted dataframe, using reshape2 <http://cran.r-project.org/web/packages/reshape2/index.html>.
The first column should contain IDs, the second column should contain variable
names, and the third column should contain variable values.}
}
\description{
Runs percent_table for melted data frames.
}

