% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/fetch_var_by.R
\name{fetch_var_by}
\alias{fetch_var_by}
\title{Fetches the values of one variable where a second variable is a certain value.}
\usage{
fetch_var_by(var, select_list, col = 3, na.rm = TRUE, data = get_data())
}
\arguments{
\item{var}{character. The name of the variable you want values returned for.}

\item{na.rm}{logical. Whether or not to remove NA and NA-like values.}

\item{select_list.}{A list of all subselections, with the name of the variable to
subselect by and the value it should be (e.g., \code{list('student' = 'Yes')})}

\item{col.}{3 (default) to grap values, 2 to grab the variable name, 1 to grab
the ids, or 'all' to grab the entire data frame for that variable.}

\item{data.}{The data to grap values from. This is assumed to be a three column
melted dataframe, using reshape2 <http://cran.r-project.org/web/packages/reshape2/index.html>.
The first column should contain IDs, the second column should contain variable
names, and the third column should contain variable values.

e.g., to return the favorite food of people from the city of Paris:
\code{fetch_var_by(favorite_food, home_city, 'Paris', data)}}
}
\description{
Fetches the values of one variable where a second variable is a certain value.
}

