% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/fetch_var_in_range.R
\name{fetch_var_in_range}
\alias{fetch_var_in_range}
\title{Fetches values from a melted data frame that are within a particular range.}
\usage{
fetch_var_in_range(var, min, max, data = get_data())
}
\arguments{
\item{var}{character. The name of the variable within the data frame to
select values from.}

\item{min}{numeric. The smaller end of the range.}

\item{max}{numeric. The upper end of the range.}

\item{data.}{The data to grap values from. This is assumed to be a three column
melted dataframe, using reshape2 <http://cran.r-project.org/web/packages/reshape2/index.html>.
The first column should contain IDs, the second column should contain variable
names, and the third column should contain variable values.}
}
\description{
Fetches values from a melted data frame that are within a particular range.
}

