% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fish_for_correlations.R
\name{fish_for_correlations}
\alias{fish_for_correlations}
\title{Fish the entire dataframe for correlations. Many will be spurious.}
\usage{
fish_for_correlations(df, x = "all", y = "all", p_threshold = 0.05)
}
\arguments{
\item{df}{dataframe. The dataframe to fish.}

\item{x}{character. A character vector of names to put first in the correlation. Or 'all' for all.}

\item{y}{character. A character vector of names to put second in the correlation. Or 'all' for all.}

\item{p_threshold}{numeric. The threshold cutoff for determining a significant test.}
}
\description{
Fish the entire dataframe for correlations. Many will be spurious.
}
