% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/fn_on_df.R
\name{fn_on_df}
\alias{fn_on_df}
\title{Runs a function on the values of a melted data frame.
Used for preforming median, mean, sd, and sum without a ton of ugly scripting.}
\usage{
fn_on_df(fetch_group, fn, ...)
}
\arguments{
\item{fetch_group.}{The result of a fetch_var()}

\item{fn.}{The name of the function.

e.g., the median age would be \code{fn_on_df(fetch('age'), median)}}
}
\description{
Runs a function on the values of a melted data frame.
Used for preforming median, mean, sd, and sum without a ton of ugly scripting.
}

