% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gather.R
\name{gather}
\alias{gather}
\title{Gets the vector of a particular variable within a dplyr postgres table.}
\usage{
gather(pt, ...)
}
\arguments{
\item{pt}{postgres.  The postgres table.}

\item{variable}{character. The name of the variable to gather.}
}
\description{
Gets the vector of a particular variable within a dplyr postgres table.
}

