% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Gets the data parameter that was previously set by set_data() or errors.}
\usage{
get_data()
}
\description{
Gets the data parameter that was previously set by set_data() or errors.
}

