% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_names.R
\name{get_names}
\alias{get_names}
\title{Gets the names of a dplyr postgres table.}
\usage{
get_names(pt)
}
\arguments{
\item{pt}{postgres.  The postgres table.}
}
\description{
Gets the names of a dplyr postgres table.
}

