% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vars.R
\name{get_vars}
\alias{get_vars}
\title{Get the names from a dataframe that match a pattern.}
\usage{
get_vars(df, pattern, collapse = FALSE, ...)
}
\arguments{
\item{df}{dataframe. The dataframe to search the names of.}

\item{pattern}{character. The pattern or patterns to search for.}

\item{collapse}{logical. When using multiple patterns, whether to return
as one list (collapse = TRUE) or as multiple lists (FALSE, default).}

\item{...}{list. Additional arguments to send to grep.}
}
\description{
Get the names from a dataframe that match a pattern.
}
