% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/not_in.R
\name{\%not_in\%}
\alias{\%not_in\%}
\title{An infix operator for determining if values are not in a list.}
\usage{
x \%not_in\% table
}
\arguments{
\item{x}{variable. The object to scan for membership.}

\item{table}{vector. A vector of objects to check against.}
}
\description{
An infix operator for determining if values are not in a list.
}
