% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/ignore_by_grep.R
\name{ignore_by_grep}
\alias{ignore_by_grep}
\title{Drops all columns that match lists of regex strings.}
\usage{
ignore_by_grep(df, ..., keep = NULL)
}
\arguments{
\item{df}{dataframe. The dataframe to drop columns from.}

\item{...}{lists. Lists containing regex strings to match against.}

\item{keep}{character. A vector of variable names to not drop, despite regex.}
}
\description{
Drops all columns that match lists of regex strings.
}

