% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na.rm.R
\name{is.na_like}
\alias{is.na_like}
\title{Determines if something is NA or NA-like (blank, the string "NA", or "N/A")}
\usage{
is.na_like(x)
}
\arguments{
\item{x.}{The vector to test.}
}
\description{
Determines if something is NA or NA-like (blank, the string "NA", or "N/A")
}
