% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/make_new_var.R
\name{make_new_var}
\alias{make_new_var}
\title{Given a melted data frame and a definition for the new variable, create
that new variable added onto the melted data frame.}
\usage{
make_new_var(new_var_name, definition, data = get_data())
}
\arguments{
\item{new_var_name}{character.  The desired name of the new variable.}

\item{definition}{numeric or character.  The values for the new variable.}

\item{data.}{The data to which the new variable will be added. This is
assumed to be a three column melted dataframe, using reshape2
<http://cran.r-project.org/web/packages/reshape2/index.html>.
The first column should contain IDs, the second column should contain variable
names, and the third column should contain variable values.}
}
\description{
Given a melted data frame and a definition for the new variable, create
that new variable added onto the melted data frame.
}

