% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/na.rm.R
\name{na.rm}
\alias{na.rm}
\title{Deletes all NA and NA-like values from a vector.}
\usage{
na.rm(vec)
}
\arguments{
\item{vec}{vector. The vector to clean.}
}
\description{
Deletes all NA and NA-like values from a vector.
}

