% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response_rate.R
\name{num_answers}
\alias{num_answers}
\title{Calculates the number of people that answered a particular question.
Assumes each column corresponds to a unique respondent.}
\usage{
num_answers(df, variable, na = NA)
}
\arguments{
\item{df}{data frame. The dataframe to count.}

\item{variable}{character. The variable to count.}

\item{na.}{NA responses in the dataframe are coded as this.}
}
\description{
Calculates the number of people that answered a particular question.
Assumes each column corresponds to a unique respondent.
}

