% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num_na.R
\name{num_na}
\alias{num_na}
\title{Calculate the number of NAs in a variable.}
\usage{
num_na(x)
}
\arguments{
\item{x}{variable. The variable to calculate the number of NAs for.}
}
\description{
Calculate the number of NAs in a variable.
}
