% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num_over_zero.R
\name{num_over_zero}
\alias{num_over_zero}
\title{Calculate the number of positive values in a variable.}
\usage{
num_over_zero(x, na.rm = FALSE)
}
\arguments{
\item{x}{numeric. The variable to calculate.}

\item{na.rm}{logical. Whether or not to remove NAs when calculating the sum.}
}
\description{
Calculate the number of positive values in a variable.
}
