% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/response_rate.R
\name{num_respondants}
\alias{num_respondants}
\title{Calculates the number of total respondents in a dataframe.
Assumes each column corresponds to a unique respondent.}
\usage{
num_respondants(df)
}
\arguments{
\item{df}{data frame. The dataframe to count.}
}
\description{
Calculates the number of total respondents in a dataframe.
Assumes each column corresponds to a unique respondent.
}

