% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response_rate.R
\name{num_respondents}
\alias{num_respondents}
\title{Calculates the number of total respondents in a dataframe.
Assumes each column corresponds to a unique respondent.}
\usage{
num_respondents(df)
}
\arguments{
\item{df}{data frame. The dataframe to count.}
}
\description{
Calculates the number of total respondents in a dataframe.
Assumes each column corresponds to a unique respondent.
}

