% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/percent_table.R
\name{percent_table}
\alias{percent_table}
\title{Prints a percent table.}
\usage{
percent_table(data, round = NULL, na.rm = TRUE)
}
\arguments{
\item{round}{integer. How many digits you want to round percentages to. Default no rounding.}

\item{na.rm}{logical. Remove all the NA, "N/A", or "" values. On by default.}

\item{var}{characer. The name of the variable you want percent breakdown for.}

\item{data.}{The data to be displayed.  This is assumed to either be a vector
or a three column melted dataframe, using reshape2 <http://cran.r-project.org/web/packages/reshape2/index.html>.}
}
\description{
Prints a percent table.
}

