% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparison_table.R
\name{print.comparison_table}
\alias{print.comparison_table}
\title{Print the table without annoyingly displaying the class.}
\usage{
\method{print}{comparison_table}(x)
}
\description{
Print the table without annoyingly displaying the class.
}

