% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/dplyr_table.R
\name{print.surveytools2_dplyr_table}
\alias{print.surveytools2_dplyr_table}
\title{Print the table without annoyingly displaying the class.}
\usage{
\method{print}{surveytools2_dplyr_table}(x)
}
\description{
Print the table without annoyingly displaying the class.
}

