% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace.R
\name{replace}
\alias{replace}
\title{Replace a value within a dataframe.}
\usage{
replace(df, col, expr, value)
}
\arguments{
\item{df}{datafarme. The dataframe to replace within.}

\item{col}{character. The name of the column to replace within.}

\item{expr}{expression. An expression of the values in the column to replace.}

\item{value}{object. The value positions where the expression is true become.}
}
\description{
Replace a value within a dataframe.
}

