% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/response_rate.R
\name{response_rate}
\alias{response_rate}
\title{Calculates the response rate to a particular question.
Assumes each column corresponds to a unique respondent.}
\usage{
response_rate(df, variable, na = NA)
}
\arguments{
\item{df}{data frame. The dataframe to count.}

\item{variable}{character. The variable to count.}

\item{na.}{NA responses in the dataframe are coded as this.}
}
\description{
Calculates the response rate to a particular question.
Assumes each column corresponds to a unique respondent.
}

