% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/set_data.R
\name{set_data}
\alias{set_data}
\title{Sets the data parameter to a certain data frame.}
\usage{
set_data(data)
}
\arguments{
\item{data.}{The data to be transformed.  This is assumed to be a three
column melted dataframe, using reshape2 <http://cran.r-project.org/web/packages/reshape2/index.html>.
The first column should contain IDs, the second column should contain variable
names, and the third column should contain variable values.}
}
\description{
Sets the data parameter to a certain data frame.
}

