% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/dplyr_table.R
\name{sort.surveytools2_dplyr_table}
\alias{sort.surveytools2_dplyr_table}
\title{Sort a dplyr table.
When both frequencies and percentages are displayed, the table can be hard to sort.}
\usage{
\method{sort}{surveytools2_dplyr_table}(x, decreasing = FALSE)
}
\description{
Sort a dplyr table.
When both frequencies and percentages are displayed, the table can be hard to sort.
}

