% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab.R
\name{sort.tab}
\alias{sort.tab}
\title{Sort a tab.
When both frequencies and percentages are displayed, the table can be hard to sort.}
\usage{
\method{sort}{tab}(x, decreasing = FALSE)
}
\description{
Sort a tab.
When both frequencies and percentages are displayed, the table can be hard to sort.
}
