% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/stattest.R
\name{stattest}
\alias{stattest}
\title{Runs a particular statistical test over a melted data frame.}
\usage{
stattest(var, by_var, test, ignore = NULL, data = get_data())
}
\arguments{
\item{var}{character. The name of the first variable in the chi-square test.}

\item{by_var}{charager. The name of the second (by) variable in the
chi-square test.}

\item{test}{character. Name of the test to run.}

\item{ignore}{vector. A vector of the names of values to be replaced by NAs
    within both the var and the by_var. Can clean up categories that are too
    small for meaningful analysis.}

\item{data.}{The data to be transformed.  This is assumed to be a three
column melted dataframe, using reshape2 <http://cran.r-project.org/web/packages/reshape2/index.html>.
The first column should contain IDs, the second column should contain variable
names, and the third column should contain variable values.}
}
\description{
Gracefully handles differences in variable size and NAs, and allows easy
removal of variables.
}

