% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_csv.R
\name{summary_csv}
\alias{summary_csv}
\title{Creates a CSV file containing summary information on the given variables.}
\usage{
summary_csv(data, filename)
}
\arguments{
\item{data}{list. A list containing all the R objects to summarize.}

\item{filename}{character. The name of the CSV file to write.}
}
\value{
TRUE. Generates a CSV file at the designated location.
}
\description{
Creates a CSV file containing summary information on the given variables.
}

