% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swap_multiple_ids.R
\name{swap_multiple_ids}
\alias{swap_multiple_ids}
\title{For all ids specified, sets their value to the assigned value within a given variable.
Like swap_by_ids, but for multiple ids and one swap value.
Useful for imputation.}
\usage{
swap_multiple_ids(df, variable, ids, value, idname = "id")
}
\arguments{
\item{df}{dataframe.  The data to be transformed.}

\item{variable}{character. The name of the variable in df where the replacement occurs.}

\item{ids}{character. A vector of ids to be transformed.}

\item{value}{character. The value to assign those ids.}

\item{idname}{character. The name of the column with ids.}
}
\description{
For all ids specified, sets their value to the assigned value within a given variable.
Like swap_by_ids, but for multiple ids and one swap value.
Useful for imputation.
}
\seealso{
swap_by_ids
}
