% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_test.R
\name{t_test}
\alias{t_test}
\title{Reformat stats::t.test to accept inputs as two vectors.}
\usage{
t_test(vector1, vector2)
}
\arguments{
\item{vector1.}{The first vector to compare.}

\item{vector2.}{The other vector to compare.}
}
\value{
t-test on the joint frequency table of the two vectors.
}
\description{
Reformat stats::t.test to accept inputs as two vectors.
}
