% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab.R
\name{tab}
\alias{tab}
\title{Makes a fancy table, inspired from `tab` from STATA.}
\usage{
tab(.data, ..., freq = TRUE, percent = FALSE, byrow = TRUE, sort = TRUE,
  sort.decreasing = TRUE, na.rm = FALSE)
}
\arguments{
\item{.data}{data frame. The data to put in the table.}

\item{...}{The variables to plot in the table. Currently supports one or two variables.}

\item{freq}{logical. If true, frequencies are printed in the table.}

\item{percent}{logical. If true, percentages are printed in the table.}

\item{byrow}{logical. If true, percentages are row percentages.  If false, column percentages.  If NULL, individual cell percentages.}

\item{sort}{logical. If true, the resulting table is sorted (if possible).}

\item{sort.decreasing}{logical. If true, the sort is preformed in a descending manner.}

\item{na.rm}{logical. If true, all NAs and NA-like entries are removed from the data before tabling.}
}
\description{
Makes a fancy table, inspired from `tab` from STATA.
}
