% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/varnames.R
\name{varnames}
\alias{varnames}
\title{Gets the variable names of the melted data frame.}
\usage{
varnames(data = get_data())
}
\arguments{
\item{data.}{The data to be transformed.  This is assumed to be a three
column melted dataframe, using reshape2 <http://cran.r-project.org/web/packages/reshape2/index.html>.
The first column should contain IDs, the second column should contain variable
names, and the third column should contain variable values.}
}
\description{
Gets the variable names of the melted data frame.
}

