% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_prefix_to_table_names.R
\name{add_prefix_to_table_names}
\alias{add_prefix_to_table_names}
\title{Adds a prefix to the names of a table.}
\usage{
add_prefix_to_table_names(df, prefix, except = NULL)
}
\arguments{
\item{df}{dataframe.  The dataframe to change the names of.}

\item{prefix}{character. The prefix to add to the names.}

\item{except}{character. A vector of variables to not apply the prefix to.}
}
\description{
Adds a prefix to the names of a table.
}

