% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/breakdown.R
\name{breakdown}
\alias{breakdown}
\title{Breakdown values of a variable by the number of people who have that value or
a higher value.}
\usage{
breakdown(var, seq, data = get_data())
}
\arguments{
\item{var}{character. The value to breakdown.}

\item{seq}{list. A sequence of values to breakdown.}

\item{data.}{The data to grap values from. This is assumed to be a three column
melted dataframe, using reshape2 <http://cran.r-project.org/web/packages/reshape2/index.html>.
The first column should contain IDs, the second column should contain variable
names, and the third column should contain variable values.

e.g., \code{breakdown('age', seq(5, 90, by = 5))} will tell me how many people
are age 5 or older, how many people are age 10 or older, how many people are
age 15 or older, etc. to age 90.}
}
\description{
Breakdown values of a variable by the number of people who have that value or
a higher value.
}

