% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/count_vars.R
\name{count_vars}
\alias{count_vars}
\title{Count the amount of "yes" responses that respondants have within a group of vars.}
\usage{
count_vars(df, vars, idname = "id", yes = 1, colname = "value")
}
\arguments{
\item{df}{dataframe.  The data to be transformed.}

\item{vars}{character. A vector of variables to check.}

\item{ids}{character. A vector of all respondent ids to return results for.}

\item{value}{character. The value to assign those ids.}
}
\description{
Count the amount of "yes" responses that respondants have within a group of vars.
}

