% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/count_vars.R
\name{count_vars}
\alias{count_vars}
\title{Count the amount of a particlar response that respondants have within a group of vars.}
\usage{
count_vars(df, vars, response = 1, idname = "id", colname = "value")
}
\arguments{
\item{df}{dataframe.  The data to be transformed.}

\item{vars}{character. A vector of variables to check.}

\item{response}{character. The value of the response that you want to count.  Defaults to 1.}

\item{idname}{character. The name of the variable in the dataframe where respondent ids are kept.  Defaults to 'id'.}

\item{colname}{character. The output will be a two-column dataframe with ids and values. The names are the idname for ids and this (default 'value') for the values.}
}
\description{
Count the amount of a particlar response that respondants have within a group of vars.
}

