% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_na_cols.R
\name{drop_na_cols}
\alias{drop_na_cols}
\title{Drops columns from the dataframe that meet a certain threshold of NAs.}
\usage{
drop_na_cols(df, threshold = 0.7)
}
\arguments{
\item{df}{dataframe.  The dataframe to drop from.}

\item{treshold}{integer.  If a column has this many percent of NAs or more, drop.}
}
\description{
Drops columns from the dataframe that meet a certain threshold of NAs.
}
