% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/gather.R
\name{gather}
\alias{gather}
\title{Gets the vector of a particular variable within a dplyr postgres table.}
\usage{
gather(pdf, ...)
}
\arguments{
\item{pdf}{data frame.  The data to be transformed.}

\item{variable}{character. The name of the variable to gather.}
}
\description{
Gets the vector of a particular variable within a dplyr postgres table.
}

