% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vars.R
\name{get_vars}
\alias{get_vars}
\title{Get the names from a dataframe that match a pattern.}
\usage{
get_vars(df, pattern, ...)
}
\arguments{
\item{df}{dataframe. The dataframe to search the names of.}

\item{pattern}{character. The pattern to search for.}

\item{...}{list. Additional arguments to send to grep.}
}
\description{
Get the names from a dataframe that match a pattern.
}
