% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/na.rm.R
\name{na.rm}
\alias{na.rm}
\title{Deletes all NA and NA-like values from a melted data frame.}
\usage{
na.rm(data)
}
\arguments{
\item{data.}{The data to be transformed.  This is assumed to either be a vector
or a three column melted dataframe, using reshape2 <http://cran.r-project.org/web/packages/reshape2/index.html>.
The first column should contain IDs, the second column should contain variable
names, and the third column should contain variable values.}
}
\description{
Deletes all NA and NA-like values from a melted data frame.
}

