% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/swap_by_ids.R
\name{swap_by_ids}
\alias{swap_by_ids}
\title{Renames values within a dataframe based on ids.
Useful for imputation.}
\usage{
swap_by_ids(df, variable, swap_list)
}
\arguments{
\item{df}{data frame.  The data to be transformed.}

\item{variable}{character. The name of the variable in df where the replacement occurs.}

\item{swap_list}{list. A list mapping ids to the desired value for that id.
e.g., to replace the value for id #233 with "cat", use
\code{list(233 = 'cat')}}
}
\description{
Renames values within a dataframe based on ids.
Useful for imputation.
}

