% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/swap_by_ids.R
\name{swap_by_ids}
\alias{swap_by_ids}
\title{Given a melted data frame, a list of ids mapping to what their value should
be for a given variable, and the name of that variable in the dataframe,
perform the desired translation.  Useful in interpolation.}
\usage{
swap_by_ids(swap_list, variable, data = get_data())
}
\arguments{
\item{swap_list}{list. A list mapping ids to the value that id should have
e.g., \code{list(1 = 'cat', 2 = 'dog', 3 = 'dog')}}

\item{variable}{character. The name of the variable where the replacement occurs.}

\item{data.}{The data to be transformed.  This is assumed to be a three
column melted dataframe, using reshape2 <http://cran.r-project.org/web/packages/reshape2/index.html>.
The first column should contain IDs, the second column should contain variable
names, and the third column should contain variable values.}
}
\description{
Given a melted data frame, a list of ids mapping to what their value should
be for a given variable, and the name of that variable in the dataframe,
perform the desired translation.  Useful in interpolation.
}

