% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/swap_by_value.R
\name{swap_by_value}
\alias{swap_by_value}
\title{Given a melted data frame, a list of values as they currently are in the
data frame mapped to a list of what you want the values to be, this method
will swap the values.  Useful for shortening up or cleaning data.}
\usage{
swap_by_value(swap_list, variable, data = get_data())
}
\arguments{
\item{swap_list}{list. A list mapping current values to what they should be.
e.g., to replace all instances of "dog" witih "cat", use
\code{list('dog' = 'cat')}}

\item{variable}{character. The name of the variable where the replacement occurs.}

\item{data.}{The data to be transformed.  This is assumed to be a three
column melted dataframe, using reshape2 <http://cran.r-project.org/web/packages/reshape2/index.html>.
The first column should contain IDs, the second column should contain variable
names, and the third column should contain variable values.}
}
\description{
Given a melted data frame, a list of values as they currently are in the
data frame mapped to a list of what you want the values to be, this method
will swap the values.  Useful for shortening up or cleaning data.
}

