% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/swap_multiple_ids.R
\name{swap_multiple_ids}
\alias{swap_multiple_ids}
\title{Given a melted data frame, a vector of ids, a value, and a variable,
this will make the value of that variable at all those ids into the given value.
Useful in interpolation.}
\usage{
swap_multiple_ids(ids, value, variable, data = get_data())
}
\arguments{
\item{ids}{vector. A list of ids to swap.}

\item{value}{character. The value that the variable will be set to.}

\item{variable}{character. The variable in which the transformations will take place.}

\item{data.}{The data to be transformed.  This is assumed to be a three
column melted dataframe, using reshape2 <http://cran.r-project.org/web/packages/reshape2/index.html>.
The first column should contain IDs, the second column should contain variable
names, and the third column should contain variable values.}
}
\description{
Given a melted data frame, a vector of ids, a value, and a variable,
this will make the value of that variable at all those ids into the given value.
Useful in interpolation.
}

