% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_test.R
\name{t_test}
\alias{t_test}
\title{Reformat stats::t.test to accept inputs as two vectors.}
\usage{
t_test(variable, groupby)
}
\arguments{
\item{variable.}{A numeric vector to compare means.}

\item{groupby.}{A vector to group by.}
}
\value{
t-test on the joint frequency table of the two vectors.
}
\description{
Reformat stats::t.test to accept inputs as two vectors.
}
