% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab.R
\name{tab}
\alias{tab}
\title{Makes a fancy table, inspired from `tab` from STATA.}
\usage{
tab(.data, ..., top = 0, freq = TRUE, percent = FALSE, byrow = TRUE,
  sort = TRUE, sort.decreasing = TRUE, na.rm = FALSE)
}
\arguments{
\item{.data}{data frame. The data to put in the table.}

\item{...}{The variables to plot in the table, plus any relevant filters contained within `filters()`. Currently supports one or two variables.}

\item{top}{numeric. If >0, all vectors will only print the top N elements and all other elements will be called "Other", where N is the value for top.}

\item{freq}{logical. If TRUE, frequencies are printed in the table.}

\item{percent}{logical. If TRUE, percentages are printed in the table.}

\item{byrow}{logical. If TRUE, percentages are row percentages.  If false, column percentages.  If NULL, individual cell percentages.}

\item{sort}{logical. If TRUE, the resulting table is sorted (if possible).}

\item{sort.decreasing}{logical. If TRUE, the sort is preformed in a descending manner.}

\item{na.rm}{logical. If TRUE, all NAs and NA-like entries are removed from the data before tabling.}
}
\description{
Makes a fancy table, inspired from `tab` from STATA.
}
