% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_summary.R
\name{var_summary}
\alias{var_summary}
\title{Prints a summary of the variable.}
\usage{
var_summary(variable, na.rm = TRUE, serialize = FALSE)
}
\arguments{
\item{variable}{character. The variable to summarize.}

\item{na.rm}{logical. True to not return NAs (default), false to return NAs.}

\item{serialize}{logical. True to join lists into serialized strings, false to not (default)}
}
\value{
the class and size of the variable.
   The mean, median, min, max, and SD of the variable if it is numeric.
   A table, head, and tail of the variable if it is not numeric.
}
\description{
Prints a summary of the variable.
}
